Dissolve mixed extension maintenance ownership
